pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spring-cloud-config-client"
version = "0.3.2"
description = "A Python library to fetch configuration properties from the Spring Cloud Config Server"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Emre Akkaya", email = "emrekaganakkaya@gmail.com" },
]
keywords = [
    "spring",
    "configuration-management",
]
dependencies = [
    "requests>=2.31.0,<3",
    "pyyaml>=6.0.1,<7",
]

[project.urls]
Homepage = "https://github.com:emrekgn/spring-cloud-config-python-client"
Repository = "https://github.com:emrekgn/spring-cloud-config-python-client"

# Ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 100
unsafe-fixes = true

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",    # isort
    "N",    # pep8-naming
    "C4",   # flake8-comprehensions
    "ICN",  # flake8-import-conventions
    "PIE",  # flake8-pie
    "T20",  # flake8-print
    "PYI",  # flake8-pyi
    "PT",   # flake8-pytest-style
    "Q",    # flake8-quotes
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "SLF",  # flake8-self
    "SLOT", # flake8-slots
    "TID",  # flake8-tidy-imports
    "TCH",  # flake8-type-checking
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "ERA",  # eradicate
    "PGH",  # pygrep-hooks
    "PL",   # Pylint
    "TRY",  # tryceratops
    "FLY",  # flynt
    "NPY",  # NumPy-specific rules
    "PERF", # Perflint
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",    # Line too long (handled by formatter)
    "B008",    # Do not perform function calls in argument defaults
    "B905",    # zip() without an explicit strict= parameter
    "C901",    # Too complex (let's be more lenient)
    "PLR0913", # Too many arguments to function call
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used in comparison
    "TRY003",  # Avoid specifying long messages outside the exception class
    "PT004",   # Fixture does not return anything, add leading underscore
    "PT011",   # pytest.raises() is too broad
    "PT012",   # pytest.raises() block should contain a single simple statement
    "N803",    # Argument names should be all lowercase
    "PGH004",  # Use specific rule codes when using `noqa`
    "I001",    # Import block is un-sorted or un-formatted
    "T201",    # print statement
    "RUF010",  # Use explicit conversion flag in formatted string
    "RUF022",  # __all__ is not sorted
    "PT019",   # Fixture without a value is injected as parameter
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "*.pyi",
]
per-file-ignores = {}
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = 80
